refactor: modularise few comps.
