+++ .env.example
@@ -1,0 +1,8 @@
+# OpenAI API (Optional - for server-side fallback when users don't use BYOK)
+OPENAI_API_KEY=sk-...
+OPENAI_MODEL=model-id
+
+# Qdrant Vector Database (for semantic caching)
+QDRANT_URL=http://localhost:6333
+EMBEDDING_MODEL=embedding-model-id
+CACHE_COLLECTION_NAME=
