+++ app/page.tsx
@@ -4,26 +4,12 @@ import { useState, useRef } from "react";
 import { useChat } from "@/hooks/useChat";
 import { ChatContainer } from "@/components/chat/chatContainer";
 import { ChatInput } from "@/components/chat/chatInput";
-import { Button } from "@/components/ui/button";
-import { ThemeToggle } from "@/components/themeToggle";
-import {
-  AlertDialog,
-  AlertDialogAction,
-  AlertDialogCancel,
-  AlertDialogContent,
-  AlertDialogDescription,
-  AlertDialogFooter,
-  AlertDialogHeader,
-  AlertDialogTitle,
-  AlertDialogTrigger,
-} from "@/components/ui/alertDialog";
-import { Plus } from "lucide-react";
-import { BYOK } from "@/components/byok";
+import { ChatHeader } from "@/components/chatHeader";
+import { EmptyState } from "@/components/emptyState";
 import { toast } from "sonner";
 
 export default function Home() {
   const { messages, isLoading, sendMessage, stopGeneration, clearChat } = useChat();
-  const [dialogOpen, setDialogOpen] = useState(false);
   const [isConfigured, setIsConfigured] = useState(false);
   const byokTriggerRef = useRef<HTMLButtonElement>(null);
 
@@ -40,27 +26,19 @@ export default function Home() {
     await sendMessage(content);
   };
 
-  function handleNewChat() {
-    clearChat();
-    setDialogOpen(false);
-  }
-
   if (!hasMessages) {
     return (
       <>
-        <div className="fixed top-4 right-4 z-50 flex items-center gap-2">
-          <BYOK 
-            autoOpen={true} 
-            onConfigured={setIsConfigured}
-            triggerRef={byokTriggerRef}
-          />
-          <ThemeToggle />
-        </div>
-        <ChatInput
+        <ChatHeader 
+          onConfigured={setIsConfigured}
+          onNewChat={clearChat}
+          byokTriggerRef={byokTriggerRef}
+          autoOpenByok={true}
+        />
+        <EmptyState
           onSend={handleSendMessage}
           isLoading={isLoading}
           onStop={stopGeneration}
-          centered
         />
       </>
     );
@@ -68,40 +46,12 @@ export default function Home() {
 
   return (
     <div className="flex h-screen flex-col">
-      <div className="fixed top-4 right-4 z-50 flex items-center gap-2">
-        <BYOK 
-          onConfigured={setIsConfigured}
-          triggerRef={byokTriggerRef}
-        />
-        <AlertDialog open={dialogOpen} onOpenChange={setDialogOpen}>
-          <AlertDialogTrigger asChild>
-            <Button
-              variant="outline"
-              size="sm"
-              className="gap-2 rounded-xl bg-background/80 backdrop-blur-sm shadow-sm hover:shadow-md transition-all"
-            >
-              <Plus className="size-4" />
-              <span className="hidden sm:inline">New Chat</span>
-            </Button>
-          </AlertDialogTrigger>
-          <AlertDialogContent className="border border-border/40">
-            <AlertDialogHeader>
-              <AlertDialogTitle>Start a new chat?</AlertDialogTitle>
-              <AlertDialogDescription>
-                This will clear your current conversation. This action cannot be undone.
-              </AlertDialogDescription>
-            </AlertDialogHeader>
-            <AlertDialogFooter>
-              <AlertDialogCancel>Cancel</AlertDialogCancel>
-              <AlertDialogAction onClick={handleNewChat}>
-                Continue
-              </AlertDialogAction>
-            </AlertDialogFooter>
-          </AlertDialogContent>
-        </AlertDialog>
-        <ThemeToggle />
-      </div>
-
+      <ChatHeader 
+        onConfigured={setIsConfigured}
+        onNewChat={clearChat}
+        showNewChat={true}
+        byokTriggerRef={byokTriggerRef}
+      />
       <ChatContainer messages={messages} isLoading={isLoading} />
       <ChatInput
         onSend={handleSendMessage}
